refactor(vis): simplify - remove dead code, merge security files
Dead exports removed:
- output.ts: removed printBanner, printStartup, separator, getVersion,
  pass, fail, badge (kept blue/gray as private - used internally by
  info/note prefixes). -80 lines
- upgrade-check.ts: removed fetchLatestVersion, readCache, writeCache,
  showUpgradeNotice, UpgradeCheckCache type from public API
- pm-runner.ts: removed runResolved (unused)

Merged files:
- security.ts + script-security.ts → security.ts (single file)
  Both dealt with supply chain security. Now one coherent module with
  sections: config checking, enforcement, native sync, script runner.
  Simplified detectYarnVersion → isYarnBerry (boolean, not version string)

Updated tests to match simplified API (output.test.ts -100 lines)
