+++ packages/tooling/vis/__tests__/script-security.test.ts
@@ -4,7 +4,7 @@ import { join } from "node:path";
 
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 
-import { enforceScriptSecurity, syncAllowBuildsToNativeConfig } from "../src/script-security";
+import { enforceScriptSecurity, syncAllowBuildsToNativeConfig } from "../src/security";
 import type { VisConfig } from "../src/workspace";
 
 // ── enforceScriptSecurity ────────────────────────────────────────────
