+++ packages/tooling/vis/src/pm-runner.ts
@@ -199,7 +199,6 @@ export {
     runOutdated,
     runPmSubcommand,
     runRemove,
-    runResolved,
     runUnlink,
     runWhy,
 };
